+++ src/pages/public/ErrorPage.jsx
@@ -1,4 +1,4 @@
-import MainNavigation from "../../components/MainNavigation";
+import MainNavigation from "../../components/navigation/MainNavigation";
 
 function ErrorPage() {
   return (
