+++ src/pages/public/PasswordReoveryPage.jsx
@@ -1,4 +1,4 @@
-import Card from "../../components/Card";
+import Card from "../../components/ui/Card";
 
 function PasswordRecoveryPage() {
   return (
